functions/update_generate_transcript_status/requirements.txt
aws-psycopg2==1.3.8
aws-lambda-powertools==2.19.0
